icn_zk

Module devtools

Functions

count_constraints

Count the number of constraints for the provided circuit.

log_constraints

Log information about all constraints in the provided circuit.

print_cs_stats

Print basic statistics about a synthesized constraint system.